2.0.7

+++ assets/lib/MODxAPI/autoTable.abstract.php
@@ -4,19 +4,23 @@ require_once('MODx.php');
 abstract class autoTable extends MODxAPI
 {
     protected $table = null;
-
+	protected $generateField = false;
+	
     public function __construct($modx, $debug = false)
     {
         parent::__construct($modx, $debug);
-        $data = $this->modx->db->getTableMetaData($this->makeTable($this->table));
-        foreach ($data as $item) {
-            if (empty($this->pkName) && $item['Key'] == 'PRI') {
-                $this->pkName = $item['Field'];
-            }
-            if ($this->pkName != $item['Field']) {
-                $this->default_field[$item['Field']] = $item['Default'];
-            }
-        }
+		if(empty($this->default_field)){
+			$data = $this->modx->db->getTableMetaData($this->makeTable($this->table));
+			foreach ($data as $item) {
+				if (empty($this->pkName) && $item['Key'] == 'PRI') {
+					$this->pkName = $item['Field'];
+				}
+				if ($this->pkName != $item['Field']) {
+					$this->default_field[$item['Field']] = $item['Default'];
+				}
+			}
+			$this->generateField = true;
+		}
     }
 
     public function edit($id)
@@ -27,10 +31,12 @@ abstract class autoTable extends MODxAPI
             $this->field = array();
             $this->set = array();
 
-            $result = $this->query("SELECT * from {$this->makeTable($this->table)} where `" . $this->pkName . "`=" . (int)$id);
-            $this->fromArray($this->modx->db->getRow($result));
-            $this->id = isset($this->field[$this->pkName]) ? $this->field[$this->pkName] : null;
-            unset($this->field[$this->pkName]);
+            $result = $this->query("SELECT * from {$this->makeTable($this->table)} where `" . $this->pkName . "`='" . $this->modx->db->escape($id)."'");
+			$this->fromArray($this->modx->db->getRow($result));
+            $this->id = $this->eraseField($this->pkName);
+			if(is_bool($this->id) && $this->id === false){
+				$this->id = null;
+			}
         }
         return $this;
     }
@@ -38,24 +44,25 @@ abstract class autoTable extends MODxAPI
     public function save($fire_events = null, $clearCache = false)
     {
         $fld = $this->toArray();
-
         foreach ($this->default_field as $key => $value) {
             if ($this->newDoc && $this->get($key) === null && $this->get($key) !== $value) {
                 $this->set($key, $value);
             }
-            $this->Uset($key);
+            if((!$this->generateField || isset($fld[$key])) && $this->get($key) !== null){
+				$this->Uset($key);
+			}
             unset($fld[$key]);
         }
         if (!empty($this->set)) {
             if ($this->newDoc) {
                 $SQL = "INSERT into {$this->makeTable($this->table)} SET " . implode(', ', $this->set);
             } else {
-                $SQL = "UPDATE {$this->makeTable($this->table)} SET " . implode(', ', $this->set) . " WHERE `" . $this->pkName . "` = " . $this->id;
+				$SQL = ($this->getID() === null) ? null : "UPDATE {$this->makeTable($this->table)} SET " . implode(', ', $this->set) . " WHERE `" . $this->pkName . "` = " . $this->getID();
             }
             $this->query($SQL);
         }
 
-        if ($this->newDoc) $this->id = $this->modx->db->getInsertId();
+        if ($this->newDoc && !empty($SQL)) $this->id = $this->modx->db->getInsertId();
         if ($clearCache) {
             $this->clearCache($fire_events);
         }
@@ -68,7 +75,9 @@ abstract class autoTable extends MODxAPI
         try {
             if (is_array($_ids) && $_ids != array()) {
                 $id = $this->sanitarIn($_ids);
-                $this->query("DELETE from {$this->makeTable($this->table)} where `" . $this->pkName . "` IN ({$id})");
+				if(!empty($id)){
+					$this->query("DELETE from {$this->makeTable($this->table)} where `" . $this->pkName . "` IN ({$id})");
+				}
                 $this->clearCache($fire_events);
             } else throw new Exception('Invalid IDs list for delete: <pre>' . print_r($ids, 1) . '</pre>');
         } catch (Exception $e) {
